+++ modules/laptop.nix
@@ -59,7 +59,7 @@
     time.timeZone = lib.mkForce null;
 
     environment.systemPackages =
-      builtins.attrValues { inherit (pkgs) aldente; };
+      builtins.attrValues { inherit (pkgs) aldente ice-bar; };
 
     launchd.user.agents.alDente = {
       command = ''"/Applications/Nix Apps/AlDente.app/Contents/MacOS/AlDente"'';